use embedme to generate markdown code blocks from example source code
- prevents drift for when someone updates example code but not the markdown docs pages that contains the same code
