+++ examples/verification-keys/package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@types/node": "^24.10.1",
+    "embedme": "^1.22.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.3"
   },
